[Bronze V] Title: 이상한 기호, Time: 212 ms, Memory: 17716 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -4,14 +4,14 @@ public class Main {
     public static void main(String[] args) {
     	Scanner sc = new Scanner(System.in);
     	
-    	int A = sc.nextInt();
-    	int B = sc.nextInt();
+    	long A = sc.nextInt();
+    	long B = sc.nextInt();
     	
-    	int result = atMark(A, B);
+    	long result = atMark(A, B);
     	System.out.println(result);
     }
     
-    public static int atMark(int A, int B) {
+    public static long atMark(long A, long B) {
     	return (A+B)*(A-B);
     }
 }
